lib/systems: Add eachSystem
This is somewhat similar to `forAllSystems` from flake.nix in that it
applies a function taking the system as argument to all systems and
returns the folded set.

The disadvantage of `forAllSystems` is that it still requires more
syntactic sugar to avoid duplication when most parts are almost
identical between systems.
Consider for example expressing the following attribute set:
```
{
	packages = {
		x86_64-linux = { ...; };
		aarch64-linux = { ...; };
		...
	}
	checks = {
		x86_64-linux = { ...; };
		aarch64-linux = { ...; };
		...
	}
}
```

When using only `forAllSystems`, we have to call it an additional time
for every attribute that we want to add for every system, i.e.:
```
{
	packages = forAllSystems (system:
		...
	)
	checks = forAllSystems (system:
		...
	)
}
```

While this makes `forAllSystems` straight forward to understand, this
can quickly get cumbersome if the number of attributes increases.

With the introduction of `eachSystem` this is reduced to:
```
eachSystem flakeExposed (system:
	{
		packages = ...;
		checks = ...;
	}
)
```

It essentially turns the attribute set inside out, so that global
iteration over the system is possible.

This is practically identical to what the popular
`flake-utils.lib.eachDefaultSystem` function is doing [0].

[0] https://github.com/numtide/flake-utils/blob/b1d9ab70662946ef0850d488da1c9019f3a9752a/lib.nix#L31-L55